@@ -1,6 +1,6 @@
 {
   "name": "beol",
-  "version": "1.0.0",
+  "version": "1.5.0",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
@@ -14,44 +14,45 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.8",
-    "@angular/cdk": "^7.3.3",
-    "@angular/common": "^7.2.8",
-    "@angular/compiler": "^7.2.8",
-    "@angular/core": "^7.2.8",
-    "@angular/flex-layout": "^7.0.0-beta.23",
-    "@angular/forms": "^7.2.8",
-    "@angular/material": "^7.3.3",
-    "@angular/platform-browser": "^7.2.8",
-    "@angular/platform-browser-dynamic": "^7.2.8",
-    "@angular/router": "^7.2.8",
-    "@knora/action": "8.0.0",
-    "@knora/core": "8.0.0",
-    "@knora/search": "8.0.0",
-    "@knora/viewer": "8.0.0",
+    "@angular/animations": "^8.0.3",
+    "@angular/cdk": "^8.0.1",
+    "@angular/common": "^8.0.3",
+    "@angular/compiler": "^8.0.3",
+    "@angular/core": "^8.0.3",
+    "@angular/flex-layout": "^8.0.0-beta.26",
+    "@angular/forms": "^8.0.3",
+    "@angular/material": "^8.1.4",
+    "@angular/platform-browser": "^8.0.3",
+    "@angular/platform-browser-dynamic": "^8.0.3",
+    "@angular/router": "^8.0.3",
+    "@knora/action": "^9.0.0",
+    "@knora/core": "^9.0.0",
+    "@knora/search": "^9.0.0",
+    "@knora/viewer": "^9.0.0",
     "core-js": "^2.5.4",
     "hammerjs": "^2.0.8",
     "jdnconvertiblecalendar": "^0.0.3",
     "jdnconvertiblecalendardateadapter": "^0.0.9",
     "json2typescript": "^1.0.6",
     "jsonld": "^1.0.4",
-    "ngx-infinite-scroll": "^6.0.1",
+    "ngx-infinite-scroll": "^8.0.0",
+    "node-sass": "^4.9.3",
     "openseadragon": "^2.4.0",
-    "rxjs": "^6.2.2",
+    "rxjs": "^6.5.2",
     "svg-overlay": "github:openseadragon/svg-overlay",
     "ts-md5": "^1.2.4",
     "tslib": "^1.9.0",
-    "zone.js": "^0.8.26"
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.13.0",
-    "@angular/cli": "~7.3.5",
-    "@angular/compiler-cli": "^7.2.8",
-    "@angular/language-service": "^7.2.8",
+    "@angular-devkit/build-angular": "~0.803.2",
+    "@angular/cli": "~8.3.2",
+    "@angular/compiler-cli": "^8.0.3",
+    "@angular/language-service": "^8.0.3",
     "@types/jasmine": "~2.8.6",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~8.9.4",
-    "codelyzer": "~4.5.0",
+    "codelyzer": "^5.0.1",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~4.0.0",
@@ -62,6 +63,6 @@
     "protractor": "~5.3.0",
     "ts-node": "~5.0.1",
     "tslint": "~5.9.1",
-    "typescript": "~3.2.4"
+    "typescript": "~3.5.3"
   }
 }
